Extend BAM to allow unlimited values for most core fields
Define a method of variable-length encoding and use it to extend
most of the core fields in BAM.  This removes the limits on
reference length, sequence length and number of CIGAR operations
that existed in the original format.  Methods are defined to
describe exactly which fields have been extended in the header
and alignment records.  To prevent the new encoding from being
misinterpreted by older BAM-reading software, the file magic
number is changed to BAM\2 where these extensions are used.

If no extensions are needed, the resulting file will be identical
to an ordinary BAM file, apart from the changed magic number.
A way of reverting such files to BAM version 1 by rewriting the
magic number is described.  It also makes it possible to
minimally support the format by simply recognising the BAM\2
magic number.

Includes guide to which extended fields need to be supported
to enable features like more CIGAR operations, long reference
sequences, and so on.