Improve description of ChaCha20 encryption
Use KiB instead of K to avoid 1000/1024 abiguity.

Increment nonce _after_ each segment.

Describe initial counter value to use when encoding/decoding a
block.  This is 1, which seems to be what libsodium uses.

Describe how to decode a block starting at a given plain-text
position.

Remove padding of partial blocks.  It's not needed just for the
end, and lacked any way of saying how  much padding was used.

Add requirement for a properly random nonce.

Change the nonce in the header to 96-bit little-endian.  This
is exactly the right length for chacha20_ietf_poly1305 and
makes more sense as a little-endian value.

Removed big-endian definition in byte ordering section, as
no longer needed.

Removed some trailing spaces.