crypt4gh: Only allow one writer's public key in the header
Move the header data encryption method and writer's public key
fields from the header packet structure to the common part of
the header at the beginning.

Bumps version number to 2, as this changes the layout of the
header in a way that is not backwards-compatible.

This closes a loop-hole	that allowed extra header packets to be
added to an existing header while only having knowledge of the
reader's public key by generating a new writer's key and using
that to encrypt the new packet.  As all header packets must now
share the same writer's public key, only holders of the
corresponding secret key can generate new header packets.